tool out spaces, chaged int to size_t

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,5 +1,3 @@
-#include <stdlib.h>
-#include <string.h>
 #include <stdio.h>
 #include "lists.h"
 /**
@@ -9,7 +7,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int count;
+	size_t count;
 
 	count = 0;
 	while (h != NULL)
@@ -18,7 +16,6 @@ size_t print_list(const list_t *h)
 		{
 			printf("[0] (nil)\n");
 		}
-
 		else
 		{
 			printf("[%d] ", h->len);
@@ -28,5 +25,5 @@ size_t print_list(const list_t *h)
 		h = h->next;
 		count++;
 	}
-return (count);
+	return (count);
 }
